Changed addresses handler

+++ internal/endpoint/http/responses/address_response.go
@@ -1,8 +1,8 @@
 package responses
 
 type AddressResponse struct {
-	ID          int64  `json:"id"`
-	UserID      int64  `json:"userID"`
+	ID          int    `json:"id"`
+	UserID      int    `json:"userID"`
 	IsActive    bool   `json:"isActive"`
 	City        string `json:"city"`
 	Street      string `json:"street"`
